requirements.txt
mcp>=1.0.0
python-arango>=8.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
